refactor: remove wrapped component and make index page and layout rsc(#28)
