--- src/components/ChakraUI/Avatar.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-import { Avatar as A, AvatarProps, ComponentWithAs } from "@chakra-ui/react";
-
-export const Avatar: ComponentWithAs<"span", AvatarProps> = (props) => <A {...props} />;
