--- src/components/ChakraUI/Box.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-import { Box as B, ChakraComponent } from "@chakra-ui/react";
-
-export const Box: ChakraComponent<"div"> = (props) => <B {...props} />;
