--- src/components/ChakraUI/Spacer.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-import { ChakraComponent, Spacer as S } from "@chakra-ui/react";
-
-export const Spacer: ChakraComponent<"div"> = (props) => <S {...props} />;
