--- src/components/ChakraUI/Tooltip.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-import { ComponentWithAs, Tooltip as T, TooltipProps } from "@chakra-ui/react";
-
-export const Tooltip: ComponentWithAs<"div", TooltipProps> = (props) => <T {...props} />;
