+++ src/components/ChakraUI/index.ts
@@ -1,10 +1,23 @@
-export { Avatar } from "./Avatar";
-export { Box } from "./Box";
-export { Flex } from "./Flex";
-export { Heading } from "./Heading";
-export { Spacer } from "./Spacer";
-export { Tooltip } from "./Tooltip";
-export { Link } from "./Link";
-export { Breadcrumb } from "./Breadcrumb";
-export { BreadcrumbItem } from "./BreadcrumbItem";
-export { BreadcrumbLink } from "./BreadcrumbLink";
+"use client";
+
+export {
+  Avatar,
+  Box,
+  Flex,
+  Heading,
+  Spacer,
+  Tooltip,
+  Link,
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  Button,
+  Stack,
+  Text,
+  IconButton,
+  Menu,
+  MenuButton,
+  MenuItem,
+  MenuList,
+  Container,
+} from "@chakra-ui/react";
